+++ service-call-app/app/dashboard/page.tsx
@@ -12,6 +12,7 @@ import {
 } from "@/components/ui/card";
 import { revalidatePath } from "next/cache";
 import TakenBy from "./[id]/TakenBy";
+import RealTimeData from "../(Real Time Data)/RealTimeData";
 
 // Server Component for data fetching
 export default async function DashboardPage() {
@@ -37,166 +38,8 @@ export default async function DashboardPage() {
         <h2 className="text-xl md:text-2xl font-semibold">Service Calls</h2>
       </div>
 
-      {/* Card Layout for Mobile */}
-      <div className="space-y-4 md:hidden">
-        {" "}
-        {/* Hidden on medium and up */}
-        {data.reverse().map((serviceCall) => (
-          <Card key={serviceCall.id}>
-            <CardHeader>
-              <CardTitle>
-                {serviceCall.location || "Unknown Location"}
-              </CardTitle>
-              <div>{serviceCall.date?.toLocaleString() || "N/A"}</div>
-            </CardHeader>
-            <CardContent>
-              <div>
-                <strong>Who Called:</strong> {serviceCall.whoCalled || "N/A"}
-              </div>
-              <div>
-                <strong>Machine:</strong> {serviceCall.machine || "N/A"}
-              </div>
-              <div>
-                <strong>Reported Problem:</strong>{" "}
-                {serviceCall.reportedProblem || "N/A"}
-              </div>
-              <div>
-                <strong>Taken By:</strong>{" "}
-                <TakenBy
-                  id={serviceCall.id}
-                  currentTakenBy={serviceCall.takenBy!}
-                />
-              </div>
-              <div>
-                <strong>Notes:</strong>{" "}
-                <div className="py-2 px-4 text-sm text-gray-700 max-w-xs break-words">
-                  {serviceCall.notes || "N/A"}
-                </div>
-              </div>
-              <div>
-                <strong>Status:</strong>
-                <Status
-                  id={serviceCall.id}
-                  currentStatus={serviceCall.status}
-                />
-              </div>
-            </CardContent>
-            <CardFooter>
-              <Link href={`/dashboard/${serviceCall.id}/edit`} passHref>
-                <button className="text-blue-500 hover:underline">Edit</button>
-              </Link>
-            </CardFooter>
-          </Card>
-        ))}
-      </div>
-
-      {/* Table Layout for Desktop */}
-      <div className="hidden md:block">
-        {" "}
-        {/* Hidden on mobile */}
-        <table className="min-w-full bg-white border border-gray-200 rounded-lg shadow-md">
-          <thead className="bg-gray-100">
-            <tr>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Date
-              </th>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Updated At
-              </th>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Location
-              </th>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Who Called
-              </th>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Machine
-              </th>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Reported Problem
-              </th>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Status
-              </th>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Taken By
-              </th>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Notes
-              </th>
-              <th className="py-2 px-4 text-left text-sm font-semibold text-gray-600 max-w-xs">
-                Actions
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            {data.length > 0 ? (
-              data.map((serviceCall) => (
-                <tr
-                  key={serviceCall.id}
-                  className="hover:bg-gray-50 border-t border-gray-200"
-                >
-                  <td className="py-2 px-4 text-sm text-gray-700 max-w-xs">
-                    {serviceCall.date?.toLocaleString() || "N/A"}
-                  </td>
-                  <td className="py-2 px-4 text-sm text-gray-700 max-w-xs">
-                    {serviceCall.updatedAt?.toLocaleString() || "N/A"}
-                  </td>
-                  <td className="py-2 px-4 text-sm text-gray-700 max-w-xs">
-                    {serviceCall.location || "N/A"}
-                  </td>
-                  <td className="py-2 px-4 text-sm text-gray-700 max-w-xs">
-                    {serviceCall.whoCalled || "N/A"}
-                  </td>
-                  <td className="py-2 px-4 text-sm text-gray-700 max-w-xs">
-                    {serviceCall.machine || "N/A"}
-                  </td>
-                  <td className="py-2 px-4 text-sm text-gray-700 max-w-xs">
-                    {serviceCall.reportedProblem || "N/A"}
-                  </td>
-                  <td
-                    className="py-2 px-4 text-sm text-gray-700 break-words"
-                    style={{ width: "125px" }}
-                  >
-                    <Status
-                      id={serviceCall.id}
-                      currentStatus={serviceCall.status}
-                    />
-                  </td>
-                  <td
-                    className="py-2 px-4 text-sm text-gray-700 break-words"
-                    style={{ width: "125px" }}
-                  >
-                    <TakenBy
-                      id={serviceCall.id}
-                      currentTakenBy={serviceCall.takenBy!}
-                    />
-                  </td>
-                  <td className="py-2 px-4 text-sm text-gray-700 max-w-xs break-words">
-                    {serviceCall.notes || "N/A"}
-                  </td>
-                  <td className="py-2 px-4 text-sm text-gray-700 max-w-xs">
-                    <Link href={`/dashboard/${serviceCall.id}/edit`} passHref>
-                      <button className="text-blue-500 hover:underline">
-                        Edit
-                      </button>
-                    </Link>
-                  </td>
-                </tr>
-              ))
-            ) : (
-              <tr>
-                <td
-                  colSpan={10}
-                  className="py-2 px-4 text-sm text-gray-700 text-center"
-                >
-                  No service calls found
-                </td>
-              </tr>
-            )}
-          </tbody>
-        </table>
-      </div>
+      {/* Service Calls Table */}
+      <RealTimeData />
     </div>
   );
 }
